+++ Lab2/functions.h
@@ -21,6 +21,6 @@ typedef struct particle Particle;
 
 Particle* readFile(char* file_name);
 void bomb(Particle * particles, int N);
-double* bomb_parallel(Particle * particles, int N, int t);
-double* bomb_parallel2(Particle * particles, int N, int t);
-double* bomb_parallel3(Particle * particles, int N, int t);
+float* bomb_parallel(Particle * particles, int N, int t);
+float* bomb_parallel2(Particle * particles, int N, int t);
+float* bomb_parallel3(Particle * particles, int N, int t);
